+++ RELEASE_NOTES.md
@@ -7,6 +7,8 @@
 - UI and performance improvements
 - New upload media added to the chat. 
 - You can now press and hold on your messages to delete your message.
+- Profile UI Updated.
+- Followers/Following list UI updated.
 
 ## Version 1.0.0
 - Initial release of SideEye